requirements.txt
torch>=2.0.0
torchvision
torchaudio
numpy
matplotlib
pyyaml
pandas
gymnasium
pytest